+++ SharePointFramework/PortfolioWebParts/src/components/PortfolioOverview/EditViewColumnsPanel/useEditViewColumnsPanel.ts
@@ -1,12 +1,11 @@
 import { IColumn } from '@fluentui/react'
-import strings from 'PortfolioWebPartsStrings'
 import _ from 'lodash'
+import { IProjectColumn } from 'pp365-shared-library'
 import { arrayMove } from 'pp365-shared-library/lib/helpers/arrayMove'
 import { useContext, useEffect, useState } from 'react'
 import { OnDragEndResponder } from 'react-beautiful-dnd'
 import { IPortfolioOverviewContext, PortfolioOverviewContext } from '../context'
 import { TOGGLE_EDIT_VIEW_COLUMNS_PANEL } from '../reducer'
-import { IProjectColumn } from 'pp365-shared-library'
 
 /**
  * Get all columns from `context.props.configuration.columns` with the selected state
@@ -67,8 +66,8 @@ export function useEditViewColumnsPanel() {
       GtPortfolioColumnOrder: JSON.stringify(columns.map((c) => c.id))
     }
 
-    await context.props.dataAdapter.updateItemInList(
-      strings.PortfolioViewsListName,
+    await context.props.dataAdapter.portalDataService.updateItemInList(
+      'PORTFOLIO_VIEWS',
       context.state.currentView.id as any,
       properties
     )
@@ -87,8 +86,8 @@ export function useEditViewColumnsPanel() {
       GtPortfolioColumnOrder: null
     }
 
-    await context.props.dataAdapter.updateItemInList(
-      strings.PortfolioViewsListName,
+    await context.props.dataAdapter.portalDataService.updateItemInList(
+      'PORTFOLIO_VIEWS',
       context.state.currentView.id as any,
       properties
     )
