+++ SharePointFramework/PortfolioWebParts/src/data/types.ts
@@ -1,8 +1,6 @@
 import { WebPartContext } from '@microsoft/sp-webpart-base'
 import {
   ItemUpdateResult,
-  QueryPropertyValueType,
-  SearchQuery,
   SearchResult,
   SortDirection
 } from '@pnp/sp'
@@ -22,51 +20,6 @@ export interface IFetchDataForViewItemResult extends SearchResult {
   SiteId: string
   [key: string]: any
 }
-
-export const DEFAULT_SEARCH_SETTINGS: SearchQuery = {
-  Querytext: '*',
-  RowLimit: 500,
-  TrimDuplicates: false,
-  Properties: [
-    {
-      Name: 'EnableDynamicGroups',
-      Value: {
-        BoolVal: true,
-        QueryPropertyValueTypeIndex: QueryPropertyValueType.BooleanType
-      }
-    }
-  ],
-  SortList: [{ Property: 'LastModifiedTime', Direction: SortDirection.Descending }]
-}
-
-export const CONTENT_TYPE_ID_BENEFITS = '0x01004F466123309D46BAB9D5C6DE89A6CF67'
-export const CONTENT_TYPE_ID_MEASUREMENTS = '0x010039EAFDC2A1624C1BA1A444FC8FE85DEC'
-export const CONTENT_TYPE_ID_INDICATORS = '0x010073043EFE3E814A2BBEF96B8457623F95'
-export const DEFAULT_GAINS_PROPERTIES = [
-  'Path',
-  'SPWebURL',
-  'Title',
-  'ListItemId',
-  'SiteTitle',
-  'SiteId',
-  'ContentTypeID',
-  'GtDesiredValueOWSNMBR',
-  'GtMeasureIndicatorOWSTEXT',
-  'GtMeasurementUnitOWSCHCS',
-  'GtStartValueOWSNMBR',
-  'GtMeasurementValueOWSNMBR',
-  'GtMeasurementCommentOWSMTXT',
-  'GtMeasurementDateOWSDATE',
-  'GtGainsResponsibleOWSUSER',
-  'GtGainsTurnoverOWSMTXT',
-  'GtGainsTypeOWSCHCS',
-  'GtPrereqProfitAchievementOWSMTXT',
-  'GtRealizationTimeOWSDATE',
-  'GtGainLookupId',
-  'GtMeasureIndicatorLookupId',
-  'GtGainsResponsible',
-  'GtGainsOwner'
-]
 
 export interface IPortfolioWebPartsDataAdapter {
   /**
@@ -305,35 +258,6 @@ export interface IPortfolioWebPartsDataAdapter {
    * @param column Column to delete
    */
   deleteProjectContentColumn?(property: Record<string, any>): Promise<any>
-
-  /**
-   * Add item to a list
-   *
-   * @param listName List name
-   * @param properties Properties
-   */
-  addItemToList?<T>(listName: string, properties: Record<string, any>): Promise<T>
-
-  /**
-   * Update item in a list
-   *
-   * @param listName List name
-   * @param itemId Item ID
-   * @param properties Properties
-   */
-  updateItemInList?<T>(
-    listName: string,
-    itemId: number,
-    properties: Record<string, any>
-  ): Promise<T>
-
-  /**
-   * Deletes the item with the specified ID from the specified list.
-   *
-   * @param listName List name
-   * @param itemId Item ID
-   */
-  deleteItemFromList?(listName: string, itemId: number): Promise<boolean>
 
   /**
    * Adds a new column to the project columns list and adds the column to the specified view.
